+++ R/extract_loci.R
@@ -1,12 +1,13 @@
 #' Extract loci from GWAS/QTL summary statistics
 #' 
 #' @param topSNPs Positions of the top SNPs prepare using  
-#' \link[echolocatoR]{import_topSNPs}.
+#' \link[echodata]{import_topSNPs}.
 #' @param fullSS \link[data.table]{data.table} of the 
 #' full GWAS/QTL summary statistics.
+#' @param bp_distance The basepair distance upstream/downstream of the proxy 
+#' SNP (in \code{topSNPs}) to extract.
 #' @param munged Whether \code{fullSS} was previously munged with 
-#' \link[MungeSumstats]{format_sumstats}.
-#' @inheritParams echolocatoR::finemap_pipeline
+#' \link[MungeSumstats]{format_sumstats}. 
 #' 
 #' @export
 #' @importFrom dplyr %>% group_by count
@@ -20,7 +21,7 @@ extract_loci <- function(topSNPs,
     # topSNPs <- echodata::topSNPs_Kunkle2019
     # fullSS <- get_Kunkle2019()
     if(munged){
-        fullSS <- MUNGESUMSTATS.to_echolocatoR(dat = fullSS)    
+        fullSS <- mungesumstats_to_echolocatoR(dat = fullSS)    
     }
     #### Subset loci ####
     locusSS <- lapply(seq(1,nrow(topSNPs)), function(i){