+++ R/mungesumstats_col_map.R
@@ -2,7 +2,7 @@
 #'
 #' @source \url{https://github.com/neurogenomics/MungeSumstats}
 #' @export
-MUNGESUMSTATS.col_map <- function(){
+mungesumstats_col_map <- function(){
     list(chrom_col="CHR",
          position_col="BP",
          snp_col="SNP",