+++ tests/testthat/test-MUNGESUMSTATS.R
@@ -1,20 +1,20 @@
-test_that("MUNGESUMSTATS functions work", {
+test_that("MungeSumstats functions work", {
   
-    #### MUNGESUMSTATS.col_map ####
-    col_map <- echodata:::MUNGESUMSTATS.col_map()
+    #### mungesumstats_col_map ####
+    col_map <- echodata:::mungesumstats_col_map()
     testthat::expect_length(col_map, 13)
     testthat::expect_true(methods::is(col_map,"list"))
     
-    #### MUNGESUMSTATS.col_map ####
+    #### mungesumstats_col_map ####
     dat <- echodata::BST1
     dat <- dplyr::rename(dat, pval = P)
-    dat2 <- echodata:::MUNGESUMSTATS.check_syn(dat = dat,
+    dat2 <- echodata:::mungesumstats_check_syn(dat = dat,
                                                col_name = "pval")
     testthat::expect_true(!"P" %in% colnames(dat))
     testthat::expect_true("P" %in% colnames(dat2))
     
     
-    #### MUNGESUMSTATS.to_echolocatoR ####
+    #### mungesumstats_to_echolocatoR ####
     # eduAttainOkbayPth <- system.file("extdata","eduAttainOkbay.txt",
     #                                  package="MungeSumstats")
     # reformatted <- MungeSumstats::format_sumstats(path=eduAttainOkbayPth,
@@ -22,7 +22,7 @@ test_that("MUNGESUMSTATS functions work", {
     #                                               return_data = TRUE)
     dat <- echodata::BST1
     dat_mss <- dplyr::rename(dat, BP=POS, FRQ=Freq, BETA=Effect)
-    dat_echoR <- echodata::MUNGESUMSTATS.to_echolocatoR(dat=dat)
+    dat_echoR <- echodata::mungesumstats_to_echolocatoR(dat=dat)
     testthat::expect_true(all(c("POS","Freq","Effect") %in% colnames(dat)))
     testthat::expect_true(!all(c("POS","Freq","Effect") %in% colnames(dat_mss)))
     testthat::expect_true(all(c("POS","Freq","Effect") %in% colnames(dat_echoR)))